First brutal pass at breaking it out from the non schema validation
