+++ tests/platform_helper/utils/test_validation.py
@@ -8,18 +8,20 @@ from schema import SchemaError
 
 from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
 from dbt_platform_helper.constants import PLATFORM_HELPER_VERSION_FILE
-from dbt_platform_helper.utils.validation import S3_BUCKET_NAME_ERROR_TEMPLATE
+from dbt_platform_helper.utils.platform_config_schema import (
+    S3_BUCKET_NAME_ERROR_TEMPLATE,
+)
+from dbt_platform_helper.utils.platform_config_schema import int_between
+from dbt_platform_helper.utils.platform_config_schema import validate_s3_bucket_name
+from dbt_platform_helper.utils.platform_config_schema import validate_string
 from dbt_platform_helper.utils.validation import _validate_extension_supported_versions
 from dbt_platform_helper.utils.validation import config_file_check
 from dbt_platform_helper.utils.validation import float_between_with_halfstep
-from dbt_platform_helper.utils.validation import int_between
 from dbt_platform_helper.utils.validation import lint_yaml_for_duplicate_keys
 from dbt_platform_helper.utils.validation import load_and_validate_platform_config
 from dbt_platform_helper.utils.validation import validate_addons
 from dbt_platform_helper.utils.validation import validate_database_copy_section
 from dbt_platform_helper.utils.validation import validate_platform_config
-from dbt_platform_helper.utils.validation import validate_s3_bucket_name
-from dbt_platform_helper.utils.validation import validate_string
 from tests.platform_helper.conftest import FIXTURES_DIR
 from tests.platform_helper.conftest import UTILS_FIXTURES_DIR
